[CELEBORN-1499] Bump Ratis version from 3.0.1 to 3.1.0
