+++ dev/deps/dependencies-server
@@ -117,15 +117,15 @@ netty-transport/4.1.109.Final//netty-transport-4.1.109.Final.jar
 osgi-resource-locator/1.0.3//osgi-resource-locator-1.0.3.jar
 paranamer/2.8//paranamer-2.8.jar
 protobuf-java/3.21.7//protobuf-java-3.21.7.jar
-ratis-client/3.0.1//ratis-client-3.0.1.jar
-ratis-common/3.0.1//ratis-common-3.0.1.jar
-ratis-grpc/3.0.1//ratis-grpc-3.0.1.jar
-ratis-metrics-default/3.0.1/ratis-metrics-default-3.0.1.jar
-ratis-netty/3.0.1//ratis-netty-3.0.1.jar
-ratis-proto/3.0.1//ratis-proto-3.0.1.jar
-ratis-server-api/3.0.1//ratis-server-api-3.0.1.jar
-ratis-server/3.0.1//ratis-server-3.0.1.jar
-ratis-shell/3.0.1//ratis-shell-3.0.1.jar
+ratis-client/3.1.0//ratis-client-3.1.0.jar
+ratis-common/3.1.0//ratis-common-3.1.0.jar
+ratis-grpc/3.1.0//ratis-grpc-3.1.0.jar
+ratis-metrics-default/3.1.0/ratis-metrics-default-3.1.0.jar
+ratis-netty/3.1.0//ratis-netty-3.1.0.jar
+ratis-proto/3.1.0//ratis-proto-3.1.0.jar
+ratis-server-api/3.1.0//ratis-server-api-3.1.0.jar
+ratis-server/3.1.0//ratis-server-3.1.0.jar
+ratis-shell/3.1.0//ratis-shell-3.1.0.jar
 ratis-thirdparty-misc/1.0.5//ratis-thirdparty-misc-1.0.5.jar
 reflections/0.10.2//reflections-0.10.2.jar
 rocksdbjni/8.11.3//rocksdbjni-8.11.3.jar
