+++ master/src/main/java/org/apache/celeborn/service/deploy/master/clustermeta/ha/StateMachine.java
@@ -45,6 +45,7 @@ import org.apache.ratis.statemachine.SnapshotInfo;
 import org.apache.ratis.statemachine.StateMachineStorage;
 import org.apache.ratis.statemachine.TransactionContext;
 import org.apache.ratis.statemachine.impl.BaseStateMachine;
+import org.apache.ratis.statemachine.impl.SimpleStateMachineStorage;
 import org.apache.ratis.statemachine.impl.SingleFileSnapshotInfo;
 import org.apache.ratis.thirdparty.com.google.protobuf.ByteString;
 import org.apache.ratis.util.ExitUtils;
@@ -60,7 +61,22 @@ import org.apache.celeborn.service.deploy.master.clustermeta.ResourceProtos.Reso
 public class StateMachine extends BaseStateMachine {
   private static final Logger LOG = LoggerFactory.getLogger(StateMachine.class);
 
-  private final CelebornStateMachineStorage storage = new CelebornStateMachineStorage();
+  private final SimpleStateMachineStorage storage =
+      new SimpleStateMachineStorage() {
+
+        File tmpDir = null;
+
+        @Override
+        public void init(RaftStorage storage) throws IOException {
+          super.init(storage);
+          tmpDir = storage.getStorageDir().getTmpDir();
+        }
+
+        @Override
+        public File getTmpDir() {
+          return tmpDir;
+        }
+      };
 
   private final HARaftServer masterRatisServer;
   private RaftGroupId raftGroupId;
