+++ project/CelebornBuild.scala
@@ -56,7 +56,7 @@ object Dependencies {
   val metricsVersion = "4.2.25"
   val mockitoVersion = "4.11.0"
   val nettyVersion = "4.1.109.Final"
-  val ratisVersion = "3.0.1"
+  val ratisVersion = "3.1.0"
   val roaringBitmapVersion = "1.0.6"
   val rocksdbJniVersion = "8.11.3"
   val jacksonVersion = "2.15.3"
